move copied files rmeoval to remotefilecopy only
