+++ src/BobAliensRecovery/AliensRecovery/BlobsMover.cs
@@ -1,8 +1,5 @@
-using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using BobAliensRecovery.AliensRecovery.Entities;
@@ -91,23 +88,7 @@ namespace BobAliensRecovery.AliensRecovery
 
             foreach (var transaction in transactions)
             {
-                var srcFiles = await _filesFinder.FindFiles(transaction.From, cancellationToken);
-                var dstFiles = await _filesFinder.FindFiles(transaction.To, cancellationToken);
-
-                var equal = srcFiles
-                    .Select(f => (transaction.From, file: f))
-                    .ToHashSet(FileInfoComparer.Instance);
-                equal.IntersectWith(dstFiles.Select(f => (transaction.To, f)));
-
-                var filesToRemove = equal.Select(t => t.file);
-
-                if (filesToRemove.Any())
-                {
-                    if (await _remoteFileCopier.RemoveFiles(filesToRemove, cancellationToken))
-                        _logger.LogInformation("Successfully removed source files from {dir}", transaction.From);
-                    else
-                        _logger.LogWarning("Failed to remove source files from {dir}", transaction.From);
-                }
+                await _remoteFileCopier.RemoveAlreadyMovedFiles(transaction.From, transaction.To, cancellationToken);
 
                 if (!cleanedUpDirectories.Contains(transaction.From))
                 {
@@ -120,23 +101,6 @@ namespace BobAliensRecovery.AliensRecovery
                         _logger.LogWarning("Failed to clean up empty directories at {dir}", transaction.From);
                 }
             }
-        }
-
-        private class FileInfoComparer : IEqualityComparer<(RemoteDir dir, RemoteFileInfo file)>
-        {
-            public bool Equals((RemoteDir dir, RemoteFileInfo file) x, (RemoteDir dir, RemoteFileInfo file) y)
-            {
-                return x.file.Checksum == y.file.Checksum
-                    && x.file.LengthBytes == y.file.LengthBytes
-                    && x.file.Filename.AsSpan(x.dir.Path.Length).SequenceEqual(y.file.Filename.AsSpan(y.dir.Path.Length));
-            }
-
-            public int GetHashCode([DisallowNull] (RemoteDir dir, RemoteFileInfo file) obj)
-            {
-                return HashCode.Combine(obj.file.Checksum);
-            }
-
-            public static FileInfoComparer Instance { get; } = new();
         }
     }
 }
